Изменен класс Settings.cs создание пользователей перенесены в классы Employee.cs и Intern.cs
