+++ Working/Person.cs
@@ -8,6 +8,11 @@ namespace Working
     {
         public int ID { get; set; } 
         public string Name { get; set; }
+
+        public Person()
+        {
+            
+        }
         
         public Person(int id, string name)
         {
